feat: cookie 인터페이스 생성 / RedirectCookie 생성
- 핸들러에서 작업 성공 시, 쿠키에 저장된 이전 요청 url으로 redirect 할 때 사용하는 RedirectCookie를 생성
